contracts: remove trait, add custom caching providers
