+++ Cargo.toml
@@ -14,7 +14,7 @@ edition = "2021"
 license = "Apache-2.0"
 
 [workspace.dependencies]
-amplify = "4.8.0"
+amplify = "4.8.1"
 strict_encoding = "2.8.1"
 strict_types = "2.8.2"
 commit_verify = "=0.12.0-beta.5"
