+++ src/contracts.rs
@@ -22,131 +22,276 @@
 // or implied. See the License for the specific language governing permissions and limitations under
 // the License.
 
-mod inmem;
-mod cached;
-
 use alloc::collections::BTreeMap;
 use core::borrow::Borrow;
+use core::cell::RefCell;
 use core::error::Error as StdError;
+use std::collections::HashMap;
 use std::io;
 
-use amplify::confinement::SmallOrdMap;
-use hypersonic::{AcceptError, Articles, CallParams, StateName, Stock};
+use amplify::confinement::{KeyedCollection, SmallOrdMap};
+use hypersonic::{
+    AcceptError, AuthToken, CallParams, CodexId, ContractId, ContractName, Opid, Schema,
+};
 use rgb::RgbSeal;
 use strict_encoding::{
     ReadRaw, StrictDecode, StrictDumb, StrictEncode, StrictReader, StrictWriter, WriteRaw,
 };
-use strict_types::StrictVal;
 
-pub use self::cached::ContractsCache;
-pub use self::inmem::ContractsInmem;
 use crate::{
-    AuthToken, CallError, CellAddr, CodexId, ConsumeError, ContractId, ContractInfo, ContractRef,
-    ContractState, CreateParams, Operation, Opid, Pile, Schema, WitnessStatus,
+    Articles, CallError, ConsumeError, Contract, ContractInfo, ContractRef, ContractState,
+    CreateParams, Operation, Pile, Stockpile, WitnessStatus,
 };
 
-#[derive(Clone, PartialEq, Eq, Debug)]
-#[cfg_attr(feature = "serde", derive(Serialize, Deserialize), serde(rename_all = "camelCase"))]
-pub struct OpOut<Seal: RgbSeal> {
-    pub addr: CellAddr,
-    pub seal: Seal,
-    pub name: StateName,
-    pub val: StrictVal,
+/// Collection of RGB smart contracts and contract issuers, which can be cached in memory.
+///
+/// # Generics
+///
+/// - `S` provides a specific cache implementation for an in-mem copy of issuers,
+/// - `C` provides a specific cache implementation for an in-mem copy of contracts.
+pub struct Contracts<
+    Sp,
+    S = HashMap<CodexId, Schema>,
+    C = HashMap<ContractId, Contract<<Sp as Stockpile>::Stock, <Sp as Stockpile>::Pile>>,
+> where
+    Sp: Stockpile,
+    S: KeyedCollection<Key = CodexId, Value = Schema>,
+    C: KeyedCollection<Key = ContractId, Value = Contract<Sp::Stock, Sp::Pile>>,
+{
+    schemata: RefCell<S>,
+    contracts: RefCell<C>,
+    persistence: Sp,
 }
 
-pub trait ContractsApi<S: Stock, P: Pile> {
-    fn codex_ids(&self) -> impl Iterator<Item = CodexId>;
-    fn issuers_count(&self) -> usize;
-    fn issuers(&self) -> impl Iterator<Item = (CodexId, Schema)>;
-    fn issuer(&self, codex_id: CodexId) -> Option<Schema>;
-
-    fn contracts_count(&self) -> usize;
-    fn contract_ids(&self) -> impl Iterator<Item = ContractId>;
-    fn contracts_info(&self) -> impl Iterator<Item = ContractInfo>;
-    fn contracts_state(&self) -> impl Iterator<Item = (ContractId, ContractState<P::Seal>)> {
-        // Some implementations, for instance, doing network requests, may provide a more efficient
-        // method
-        self.contract_ids().map(|id| (id, self.contract_state(id)))
+impl<Sp, S, C> Contracts<Sp, S, C>
+where
+    Sp: Stockpile,
+    S: KeyedCollection<Key = CodexId, Value = Schema>,
+    C: KeyedCollection<Key = ContractId, Value = Contract<Sp::Stock, Sp::Pile>>,
+{
+    pub fn load(persistence: Sp) -> Self
+    where
+        S: Default,
+        C: Default,
+    {
+        Self { schemata: none!(), contracts: none!(), persistence }
     }
-    fn contract_state(&self, contract_id: ContractId) -> ContractState<P::Seal>;
-    fn contract_articles(&self, contract_id: ContractId) -> Articles;
 
-    fn has_contract(&self, contract_id: ContractId) -> bool;
-    fn find_contract_id(&self, r: impl Into<ContractRef>) -> Option<ContractId>;
+    fn with_contract<R>(
+        &self,
+        id: ContractId,
+        f: impl FnOnce(&Contract<Sp::Stock, Sp::Pile>) -> R,
+        or: Option<R>,
+    ) -> R {
+        if let Some(contract) = self.contracts.borrow().get(&id) {
+            f(contract)
+        } else if let Some(contract) = self.persistence.contract(id) {
+            let res = f(&contract);
+            self.contracts.borrow_mut().insert(id, contract);
+            res
+        } else if let Some(or) = or {
+            or
+        } else {
+            panic!("Contract {} not found", id)
+        }
+    }
 
-    /// Iterates over all witness ids known to the set of contracts.
-    ///
-    /// # Nota bene
-    ///
-    /// Iterator may repeat the same id multiple times.
-    fn witness_ids(&self) -> impl Iterator<Item = <P::Seal as RgbSeal>::WitnessId>;
-
-    fn import(&mut self, schema: Schema) -> Result<CodexId, impl StdError>;
-
-    fn issue(
+    fn with_contract_mut<R>(
         &mut self,
-        params: CreateParams<<P::Seal as RgbSeal>::Definiton>,
-    ) -> Result<ContractId, IssueError<impl StdError>>;
+        id: ContractId,
+        f: impl FnOnce(&mut Contract<Sp::Stock, Sp::Pile>) -> R,
+    ) -> R {
+        if let Some(contract) = self.contracts.borrow_mut().get_mut(&id) {
+            f(contract)
+        } else if let Some(mut contract) = self.persistence.contract(id) {
+            let res = f(&mut contract);
+            self.contracts.borrow_mut().insert(id, contract);
+            res
+        } else {
+            panic!("Contract {} not found", id)
+        }
+    }
+}
 
-    fn contract_call(
+impl<Sp, S, C> Contracts<Sp, S, C>
+where
+    Sp: Stockpile,
+    S: KeyedCollection<Key = CodexId, Value = Schema>,
+    C: KeyedCollection<Key = ContractId, Value = Contract<Sp::Stock, Sp::Pile>>,
+{
+    pub fn codex_ids(&self) -> impl Iterator<Item = CodexId> + use<'_, Sp, S, C> {
+        self.persistence.codex_ids()
+    }
+
+    pub fn issuers_count(&self) -> usize { self.persistence.issuers_count() }
+
+    pub fn issuers(&self) -> impl Iterator<Item = (CodexId, Schema)> + use<'_, Sp, S, C> {
+        self.persistence
+            .codex_ids()
+            .filter_map(|codex_id| self.issuer(codex_id).map(|schema| (codex_id, schema)))
+    }
+
+    pub fn issuer(&self, codex_id: CodexId) -> Option<Schema> {
+        if let Some(issuer) = self.schemata.borrow().get(&codex_id) {
+            return Some(issuer.clone());
+        };
+        let issuer = self.persistence.issuer(codex_id)?;
+        self.schemata.borrow_mut().insert(codex_id, issuer);
+        self.schemata.borrow().get(&codex_id).cloned()
+    }
+
+    pub fn contracts_count(&self) -> usize { self.persistence.contracts_count() }
+
+    pub fn contract_ids(&self) -> impl Iterator<Item = ContractId> + use<'_, Sp, S, C> {
+        self.persistence.contract_ids()
+    }
+
+    pub fn contracts_info(&self) -> impl Iterator<Item = ContractInfo> + use<'_, Sp, S, C> {
+        self.contract_ids().filter_map(|id| {
+            self.with_contract(
+                id,
+                |contract| Some(ContractInfo::new(id, contract.articles())),
+                Some(None),
+            )
+        })
+    }
+
+    pub fn contract_state(
+        &self,
+        contract_id: ContractId,
+    ) -> ContractState<<Sp::Pile as Pile>::Seal> {
+        self.with_contract(contract_id, |contract| contract.state(), None)
+    }
+
+    pub fn contract_articles(&self, contract_id: ContractId) -> Articles {
+        self.with_contract(contract_id, |contract| contract.articles().clone(), None)
+    }
+
+    pub fn has_contract(&self, contract_id: ContractId) -> bool {
+        self.persistence.has_contract(contract_id)
+    }
+
+    pub fn find_contract_id(&self, r: impl Into<ContractRef>) -> Option<ContractId> {
+        match r.into() {
+            ContractRef::Id(id) if self.has_contract(id) => Some(id),
+            ContractRef::Id(_) => None,
+            ContractRef::Name(name) => {
+                let name = ContractName::Named(name);
+                if let Some(id) = self
+                    .contracts
+                    .borrow()
+                    .iter()
+                    .find(|(_, contract)| contract.articles().issue.meta.name == name)
+                    .map(|(id, _)| *id)
+                {
+                    return Some(id);
+                }
+                self.persistence
+                    .contract_ids()
+                    .filter_map(|id| self.persistence.contract(id))
+                    .find(|contract| contract.articles().issue.meta.name == name)
+                    .map(|contract| contract.contract_id())
+            }
+        }
+    }
+
+    pub fn witness_ids(
+        &self,
+    ) -> impl Iterator<Item = <<Sp::Pile as Pile>::Seal as RgbSeal>::WitnessId> + use<'_, Sp, S, C>
+    {
+        self.persistence.contract_ids().flat_map(move |id| {
+            self.with_contract(
+                id,
+                |contract| contract.witness_ids().collect::<Vec<_>>(),
+                Some(none!()),
+            )
+        })
+    }
+
+    pub fn import(&mut self, schema: Schema) -> Result<CodexId, impl StdError + use<'_, Sp, S, C>> {
+        let codex_id = schema.codex.codex_id();
+        let schema = match self.persistence.import(schema) {
+            Ok(schema) => schema,
+            Err(err) => return Err(err),
+        };
+        self.schemata.borrow_mut().insert(codex_id, schema);
+        Ok(codex_id)
+    }
+
+    pub fn issue(
+        &mut self,
+        params: CreateParams<<<Sp::Pile as Pile>::Seal as RgbSeal>::Definiton>,
+    ) -> Result<ContractId, IssueError<impl StdError>> {
+        let contract = self.persistence.issue(params)?;
+        let id = contract.contract_id();
+        self.contracts.borrow_mut().insert(id, contract);
+        Ok(id)
+    }
+
+    pub fn contract_call(
         &mut self,
         contract_id: ContractId,
         call: CallParams,
-        seals: SmallOrdMap<u16, <P::Seal as RgbSeal>::Definiton>,
-    ) -> Result<Operation, AcceptError>;
+        seals: SmallOrdMap<u16, <<Sp::Pile as Pile>::Seal as RgbSeal>::Definiton>,
+    ) -> Result<Operation, AcceptError> {
+        self.with_contract_mut(contract_id, |contract| contract.call(call, seals))
+    }
 
-    fn sync(
+    pub fn sync(
         &mut self,
-        changed: impl IntoIterator<Item = (<P::Seal as RgbSeal>::WitnessId, WitnessStatus)>,
-    ) -> Result<(), AcceptError>;
+        changed: impl IntoIterator<
+            Item = (<<Sp::Pile as Pile>::Seal as RgbSeal>::WitnessId, WitnessStatus),
+        >,
+    ) -> Result<(), AcceptError> {
+        let contract_ids = self.persistence.contract_ids().collect::<Vec<_>>();
+        for (id, status) in changed {
+            for contract_id in &contract_ids {
+                self.with_contract_mut(*contract_id, |contract| contract.sync(id, status))?;
+            }
+        }
+        Ok(())
+    }
 
-    fn include(
+    pub fn include(
         &mut self,
         contract_id: ContractId,
         opid: Opid,
-        pub_witness: &<P::Seal as RgbSeal>::Published,
-        anchor: <P::Seal as RgbSeal>::Client,
-    );
+        pub_witness: &<<Sp::Pile as Pile>::Seal as RgbSeal>::Published,
+        anchor: <<Sp::Pile as Pile>::Seal as RgbSeal>::Client,
+    ) {
+        self.with_contract_mut(contract_id, |contract| contract.include(opid, anchor, pub_witness))
+    }
 
-    fn consign(
+    pub fn consign(
         &mut self,
         contract_id: ContractId,
         terminals: impl IntoIterator<Item = impl Borrow<AuthToken>>,
         writer: StrictWriter<impl WriteRaw>,
     ) -> io::Result<()>
     where
-        <P::Seal as RgbSeal>::Client: StrictDumb + StrictEncode,
-        <P::Seal as RgbSeal>::Published: StrictDumb + StrictEncode,
-        <P::Seal as RgbSeal>::WitnessId: StrictEncode;
-
-    #[cfg(feature = "fs")]
-    fn consign_to_file(
-        &mut self,
-        contract_id: ContractId,
-        terminals: impl IntoIterator<Item = impl Borrow<AuthToken>>,
-        path: impl AsRef<std::path::Path>,
-    ) -> io::Result<()>
-    where
-        <P::Seal as RgbSeal>::Client: StrictDumb + StrictEncode,
-        <P::Seal as RgbSeal>::Published: StrictDumb + StrictEncode,
-        <P::Seal as RgbSeal>::WitnessId: StrictEncode,
+        <<Sp::Pile as Pile>::Seal as RgbSeal>::Client: StrictDumb + StrictEncode,
+        <<Sp::Pile as Pile>::Seal as RgbSeal>::Published: StrictDumb + StrictEncode,
+        <<Sp::Pile as Pile>::Seal as RgbSeal>::WitnessId: StrictEncode,
     {
-        use strict_encoding::StreamWriter;
-        let file = std::fs::File::create_new(path)?;
-        let writer = StrictWriter::with(StreamWriter::new::<{ usize::MAX }>(file));
-        self.consign(contract_id, terminals, writer)
+        self.with_contract_mut(contract_id, |contract| contract.consign(terminals, writer))
     }
 
-    fn consume(
+    pub fn consume(
         &mut self,
         contract_id: ContractId,
         reader: &mut StrictReader<impl ReadRaw>,
-        seal_resolver: impl FnMut(&Operation) -> BTreeMap<u16, <P::Seal as RgbSeal>::Definiton>,
-    ) -> Result<(), ConsumeError<<P::Seal as RgbSeal>::Definiton>>
+        seal_resolver: impl FnMut(
+            &Operation,
+        )
+            -> BTreeMap<u16, <<Sp::Pile as Pile>::Seal as RgbSeal>::Definiton>,
+    ) -> Result<(), ConsumeError<<<Sp::Pile as Pile>::Seal as RgbSeal>::Definiton>>
     where
-        <P::Seal as RgbSeal>::Client: StrictDecode,
-        <P::Seal as RgbSeal>::Published: StrictDecode,
-        <P::Seal as RgbSeal>::WitnessId: StrictDecode;
+        <<Sp::Pile as Pile>::Seal as RgbSeal>::Client: StrictDecode,
+        <<Sp::Pile as Pile>::Seal as RgbSeal>::Published: StrictDecode,
+        <<Sp::Pile as Pile>::Seal as RgbSeal>::WitnessId: StrictDecode,
+    {
+        self.with_contract_mut(contract_id, |contract| contract.consume(reader, seal_resolver))
+    }
 }
 
 #[derive(Debug, Display, Error, From)]
